python/requirements.txt
langchain
matplotlib
openai
openai-whisper
pyaudio
pymysql
pyttsx3
wave
